+++ CHANGELOG.md
@@ -113,6 +113,8 @@
 *   Move `Math` and `Encoding` to stdlib. Can be required using
     `require 'math'`, etc.
 
+*   Fix some stdlib `Date` methods.
+
 ## 0.5.5 2013-11-25
 
 *   Fix regression: add `%i[foo bar]` style words back to lexer